requirements.txt
crewai==0.5.0
python-dotenv==1.0.0
requests==2.31.0
anthropic>=0.25.0
langchain-anthropic==0.1.11
beautifulsoup4
lxml
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0